Factor out Python specific build components
