+++ tooling/src/hypothesistooling/__main__.py
@@ -19,13 +19,13 @@ from __future__ import division, print_function, absolute_import
 
 import os
 import sys
-import shutil
 import subprocess
 from glob import glob
 from datetime import datetime
 
 import hypothesistooling as tools
 import hypothesistooling.installers as install
+import hypothesistooling.projects.hypothesispython as hp
 from hypothesistooling import fix_doctests as fd
 from hypothesistooling.scripts import pip_tool
 
@@ -65,74 +65,59 @@ def lint():
     )
 
 
-@task(if_changed=tools.PYTHON_SRC)
+@task(if_changed=hp.PYTHON_SRC)
 def check_type_hints():
-    pip_tool('mypy', tools.PYTHON_SRC)
+    pip_tool('mypy', hp.PYTHON_SRC)
 
 
-DIST = os.path.join(tools.HYPOTHESIS_PYTHON, 'dist')
-PENDING_STATUS = ('started', 'created')
+HEAD = tools.hash_for_name('HEAD')
+MASTER = tools.hash_for_name('origin/master')
+
+
+def do_release(package):
+    if not package.has_release():
+        print('No release for %s' % (package.__name__,))
+        return
+
+    os.chdir(package.BASE_DIR)
+
+    print('Updating changelog and version')
+    package.update_for_pending_release()
+    package.build_distribution()
+
+    print('Looks good to release!')
+
+    tag_name = package.tag_name()
+
+    print('Creating tag %s' % (tag_name,))
+
+    tools.create_tag(tag_name)
+
+    print('Uploading distribution')
+    package.upload_distribution()
+
+    print('Pushing tag')
+    tools.push_tag(tag_name)
 
 
 @task()
 def deploy():
-    os.chdir(tools.HYPOTHESIS_PYTHON)
-
-    last_release = tools.latest_version()
-
-    print('Current version: %s. Latest released version: %s' % (
-        tools.__version__, last_release
-    ))
-
-    HEAD = tools.hash_for_name('HEAD')
-    MASTER = tools.hash_for_name('origin/master')
     print('Current head:', HEAD)
     print('Current master:', MASTER)
 
-    on_master = tools.is_ancestor(HEAD, MASTER)
-    has_release = tools.has_release()
-
-    if has_release:
-        print('Updating changelog and version')
-        tools.update_for_pending_release()
-
-    print('Building an sdist...')
-
-    if os.path.exists(DIST):
-        shutil.rmtree(DIST)
-
-    subprocess.check_output([
-        sys.executable, 'setup.py', 'sdist', '--dist-dir', DIST,
-    ])
-
-    if not on_master:
+    if not tools.is_ancestor(HEAD, MASTER):
         print('Not deploying due to not being on master')
         sys.exit(0)
 
-    if not has_release:
-        print('Not deploying due to no release')
-        sys.exit(0)
-
-    print('Looks good to release!')
-
     if os.environ.get('TRAVIS_SECURE_ENV_VARS', None) != 'true':
-        print("But we don't have the keys to do it")
+        print('Running without access to secure variables, so no deployment')
         sys.exit(0)
 
     print('Decrypting secrets')
     tools.decrypt_secrets()
+    tools.configure_git()
 
-    print('Release seems good. Pushing to github now.')
-
-    tools.create_tag_and_push()
-
-    print('Now uploading to pypi.')
-
-    subprocess.check_call([
-        sys.executable, '-m', 'twine', 'upload',
-        '--config-file', tools.PYPIRC,
-        os.path.join(DIST, '*'),
-    ])
+    do_release(hp)
 
     sys.exit(0)
 
@@ -283,23 +268,12 @@ def check_requirements():
     check_not_changed()
 
 
-def update_changelog_for_docs():
-    if not tools.has_release():
-        return
-    if tools.has_uncommitted_changes(tools.CHANGELOG_FILE):
-        print(
-            'Cannot build documentation with uncommitted changes to '
-            'changelog and a pending release. Please commit your changes or '
-            'delete your release file.')
-        sys.exit(1)
-    tools.update_changelog_and_version()
-
-
-@task(if_changed=tools.HYPOTHESIS_PYTHON)
+@task(if_changed=hp.HYPOTHESIS_PYTHON)
 def documentation():
-    os.chdir(tools.HYPOTHESIS_PYTHON)
+    os.chdir(hp.HYPOTHESIS_PYTHON)
     try:
-        update_changelog_for_docs()
+        if hp.has_release():
+            hp.update_changelog_and_version()
         pip_tool(
             'sphinx-build', '-W', '-b', 'html', '-d', 'docs/_build/doctrees',
             'docs', 'docs/_build/html'
@@ -310,9 +284,9 @@ def documentation():
         ])
 
 
-@task(if_changed=tools.HYPOTHESIS_PYTHON)
+@task(if_changed=hp.HYPOTHESIS_PYTHON)
 def doctest():
-    os.chdir(tools.HYPOTHESIS_PYTHON)
+    os.chdir(hp.HYPOTHESIS_PYTHON)
     env = dict(os.environ)
     env['PYTHONPATH'] = 'src'
 
@@ -333,7 +307,7 @@ def run_tox(task, version):
     except FileExistsError:
         pass
 
-    os.chdir(tools.HYPOTHESIS_PYTHON)
+    os.chdir(hp.HYPOTHESIS_PYTHON)
     env = dict(os.environ)
     python = install.python_executable(version)
 
@@ -365,7 +339,7 @@ for n in [PY27, PY34, PY35, PY36]:
     ALIASES[n] = 'python%s.%s' % (major, minor)
 
 
-python_tests = task(if_changed=(tools.PYTHON_SRC, tools.PYTHON_TESTS))
+python_tests = task(if_changed=(hp.PYTHON_SRC, hp.PYTHON_TESTS))
 
 
 @python_tests
@@ -427,8 +401,8 @@ def check_quality():
     run_tox('quality2', PY27)
 
 
-examples_task = task(if_changed=(tools.PYTHON_SRC, os.path.join(
-    tools.HYPOTHESIS_PYTHON, 'examples')
+examples_task = task(if_changed=(hp.PYTHON_SRC, os.path.join(
+    hp.HYPOTHESIS_PYTHON, 'examples')
 ))
